Six-Hundred-Seventy-Four Commit: Refactor getTotalAmount() method to use float return type

+++ src/main/java/com/bookstore/controller/frontend/shoppingcart/ShoppingCart.java
@@ -35,8 +35,8 @@ public class ShoppingCart {
 		return total;
 	}
 	
-	public double getTotalAmount() {
-		double total = 0.0f;
+	public float getTotalAmount() {
+		float total = 0.0f;
 		Iterator<Book> iterator = cart.keySet().iterator();
 		while (iterator.hasNext()) {
 			Book book = iterator.next();
